Made CashDispenser class, changed withdraw methods
Created a CashDispenser class to house the dispensing method. The
BankData class can now talk to it when it needs math done instead of
being used by the Teller class.

Refactored the withdrawal methods do divide the workload and put
methods and instance variables in their proper places.
